+++ src/solver.cpp
@@ -2,8 +2,10 @@
 
 #include "solver.h"
 
+// todo if solver stuff stops working, try testing it with just the in data types to see if the int8_ts truncate information
+
 namespace IBN5100 {
-    int Solver::negamax(Position const &pos, int alpha, int beta) {
+    int8_t Solver::negamax(Position const &pos, int8_t alpha, int8_t beta) {
         assert(alpha < beta);
         assert(!pos.canWinNext());
 
@@ -19,7 +21,7 @@ namespace IBN5100 {
         if (moves >= 40) { return 0; }
 
         // Compute the new minimum possible score since our opponent cannot win on their next move.
-        int min = -(40 - moves)/2;
+        int8_t min = -(40 - moves)/2;
         
         // Check if we need to update our lower bound.
         if (alpha < min) {
@@ -28,7 +30,7 @@ namespace IBN5100 {
         }
 
         // Compute the maximum possible score as we cannot win on this move.
-        int max = (41 - moves)/2;
+        int8_t max = (41 - moves)/2;
 
         // Check if we need to update our upper bound.
         if (beta > max) {
@@ -42,12 +44,12 @@ namespace IBN5100 {
 
         // Simulate each possible move.
         // The max score from all the possible moves is the score of the current position.
-        for (int i = 0; i < 7; ++i) {
+        for (uint8_t i = 0; i < 7; ++i) {
             Position pos2(pos);
             pos2.play(colOrder[i]);
 
             // The score of the move would be equal to the negative score of the move for the opponent. 
-            int score = -negamax(pos2, -beta, -alpha);
+            int8_t score = -negamax(pos2, -beta, -alpha);
 
             // If the score is greater than or equal to the upper bound, we know we have found the best possible score.
             if (score >= beta) { return score; }
@@ -59,11 +61,11 @@ namespace IBN5100 {
         return alpha;
     };
 
-    int Solver::solve(Position const &pos, bool weak) {
+    int8_t Solver::solve(Position const &pos, bool weak) {
         if (pos.canWinNext()) { return (43 - pos.getMoves())/2; }
 
-        int min = -(42 - pos.getMoves())/2;
-        int max = (43 - pos.getMoves())/2 - 1; // subtract 1 as we cannot win this turn
+        int8_t min = -(42 - pos.getMoves())/2;
+        int8_t max = (43 - pos.getMoves())/2 - 1; // subtract 1 as we cannot win this turn
 
         if (weak) {
             min = -1;
@@ -72,14 +74,14 @@ namespace IBN5100 {
 
         // iteratively narrow the search window
         while (min < max) {
-            int med = min + (min + max)/2;
+            int8_t med = min + (min + max)/2;
 
             if (med <= 0 && min/2 < med) { med = min/2; }
             else if (med >= 0 && max/2 > med) { med = max/2; }
 
             // Use a search window of depth 1 to see if the actual score is less than or greater than med.
             // From this result, we can then modify the min or max accordingly.
-            int temp = negamax(pos, med, med + 1);
+            int8_t temp = negamax(pos, med, med + 1);
 
             if (temp <= med) { max = temp; }
             else { min = temp; }
